[REQUESTS] Fix bugs

+++ mpi-backend/src/main/java/itmo/mpi/repository/TripRequestRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TripRequestRepository extends JpaRepository<TripRequest, Long> {
+public interface TripRequestRepository extends JpaRepository<TripRequest, Integer> {
 }
